refactor Profile and create ProfileDetailsForm component, update profile details context, other misc changes
