+++ client/src/components/ImageUploadField/style.ts
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
-import { themeColors } from '../../styles/Theme';
+
 import { Breakpoints } from '../../styles/Breakpoints';
+import { themeColors } from '../../styles/Theme';
 
 export const StyledProfilePictureCard = styled.div`
   background-color: ${themeColors.lightGray};
