+++ client/src/components/ProfileDetailsFields/index.ts
@@ -1,0 +1,3 @@
+import ProfileDetailsFields from './ProfileDetailsFields';
+
+export default ProfileDetailsFields;
