--- client/src/components/ProfileForm/index.ts
@@ -1,3 +1,0 @@
-import ProfileForm from './ProfileForm';
-
-export default ProfileForm;
