--- client/src/components/ProfilePictureCard/index.ts
@@ -1,3 +1,0 @@
-import ProfilePictureCard from './ProfilePictureCard';
-
-export default ProfilePictureCard;
