+++ client/src/pages/Preview/Preview.tsx
@@ -1,14 +1,15 @@
 import { useContext } from 'react';
-import PlatformLink from '../../components/PlatformLink';
-import PreviewCard from '../../components/PreviewCard';
-import { LinkContext } from '../../contexts/linkContext';
-import { IShareableLinkValues } from '../../types/shareableLinkValues';
-import { platforms } from '../../utils/platformList';
+
 import {
   StyledPreview,
   StyledPreviewContainer,
   StyledPreviewLinkWrapper,
 } from './style';
+import PlatformLink from '../../components/PlatformLink';
+import PreviewCard from '../../components/PreviewCard';
+import { LinkContext } from '../../contexts/linkContext';
+import { IShareableLinkValues } from '../../types/shareableLinkValues';
+import { platforms } from '../../utils/platformList';
 
 const Preview = () => {
   const { links } = useContext(LinkContext);
