+++ client/src/types/errors.ts
@@ -3,3 +3,9 @@ export interface ILinkCardError {
   text: boolean;
   attemptedSave: boolean;
 }
+
+export type ProfileDetailsFieldsError = {
+  firstName: boolean;
+  lastName: boolean;
+  email: boolean;
+};
