+++ client/src/types/profileDetails.ts
@@ -1,11 +1,13 @@
+export type ProfilePicture = {
+  id: string;
+  name: string;
+};
+
 export interface IProfileDetails {
   firstName: string;
   lastName: string;
   email: string;
-  profilePicture: {
-    src: string;
-    name: string;
-  };
+  profilePicture: ProfilePicture;
 }
 
 export type SnackbarType = 'success' | 'error' | 'info' | 'warning';
